refactor: rename useKeyPair, remove un used import
