+++ src/components/_app/Web3ConnectionManager/components/WalletSelectorModal/components/ConnectorButton.tsx
@@ -5,9 +5,9 @@ import { Web3ReactHooks, useWeb3React } from "@web3-react/core"
 import { GnosisSafe } from "@web3-react/gnosis-safe"
 import { MetaMask } from "@web3-react/metamask"
 import { WalletConnect } from "@web3-react/walletconnect-v2"
+import useKeyPair from "components/_app/useKeyPairContext"
 import Button from "components/common/Button"
 import GuildAvatar from "components/common/GuildAvatar"
-import useKeyPair from "hooks/useKeyPair"
 import { Dispatch, SetStateAction, useMemo, useRef, useState } from "react"
 import { isMobile } from "react-device-detect"
 import { WalletError } from "types"
