+++ src/components/_app/Web3ConnectionManager/components/WalletSelectorModal/hooks/useShouldLinkToUser.ts
@@ -1,5 +1,8 @@
 import { useUserPublic } from "components/[guild]/hooks/useUser"
-import { deleteKeyPairFromIdb, getKeyPairFromIdb } from "hooks/useKeyPair"
+import {
+  deleteKeyPairFromIdb,
+  getKeyPairFromIdb,
+} from "components/_app/useKeyPairContext"
 import useSWR, { mutate, unstable_serialize } from "swr"
 import { useWeb3ConnectionManager } from "../../../Web3ConnectionManager"
