+++ src/components/_app/Web3ConnectionManager/hooks/useConnectFromLocalStorage.ts
@@ -2,7 +2,7 @@ import useUser from "components/[guild]/hooks/useUser"
 import { useConnect } from "components/[guild]/JoinModal/hooks/useConnectPlatform"
 import { Message } from "components/[guild]/JoinModal/hooks/useOauthPopupWindow"
 import { usePostHogContext } from "components/_app/PostHogProvider"
-import useKeyPair from "hooks/useKeyPair"
+import useKeyPair from "components/_app/useKeyPairContext"
 import useToast from "hooks/useToast"
 import platforms from "platforms/platforms"
 import { useEffect } from "react"
